TJShared: - Timestamp uses QueryPerformanceCounter again on Windows; there was a problem with the large numbers from Date::GetAbsoluteDate; timelines wouldn't play anymore.
TJNP:
- NetworkAddress now uses WSAAddressToString instead of inet_ntop so TJNP still works under Windows XP
- When a WebServer is created with port equal to KPortDontCare, the web server now tries to bind a socket over IPv4 first; if succeeded, the same port number is used to bind the IPv6 socket. 

TJSharedUI:
- Disabled tab appear animation in TabWnd, because it would render tabs wrongly (without text) under Windows 7 until a new tab was added.

TJShow:
- Added first version of an EP server to TJShow (using EPFramework); currently only supports global play/stop/pause/resume as commands.

git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@648 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9

+++ TJShared/include/tjtime.h
@@ -19,8 +19,8 @@ namespace tj {
 		typedef char Minutes; /* 0..59 */
 		typedef char Hours; /* 0..23 */
 
-		typedef double AbsoluteDate;
-		typedef double AbsoluteDateInterval;
+		typedef long double AbsoluteDate;
+		typedef long double AbsoluteDateInterval;
 
 		class EXPORTED Date {
 			public:
